requirements.txt
#LangChain Core
langchain
langchain-core
langchain-community

#Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai

#Enviornment Variable Management
python-dotenv

#Other utilities
fastapi      #Backend
streamlit    #Frontend
uvicorn      #FastAPI dependency
pypdf        #PDF Parsing
matplotlib   #Plotting Charts
seaborn      #Plotting Charts
python-multipart  #FastAPI dependency
chromadb     #For storing expense summary embeddings
langchain-chroma